Add JSON tests for stats APIs
