+++ pkg/controller/stats/composite.go
@@ -15,201 +15,14 @@
 package stats
 
 import (
-	"bytes"
-	"encoding/csv"
-	"encoding/json"
-	"fmt"
 	"net/http"
 	"sort"
-	"strconv"
-	"strings"
 	"time"
 
-	"github.com/google/exposure-notifications-verification-server/internal/icsv"
-	"github.com/google/exposure-notifications-verification-server/internal/project"
 	"github.com/google/exposure-notifications-verification-server/pkg/controller"
 	"github.com/google/exposure-notifications-verification-server/pkg/database"
 	"github.com/google/exposure-notifications-verification-server/pkg/rbac"
-
-	keyserver "github.com/google/exposure-notifications-server/pkg/api/v1"
 )
-
-// Type assertions for CompositeStats
-var (
-	_ icsv.Marshaler = (CompositeStats)(nil)
-	_ json.Marshaler = (CompositeStats)(nil)
-)
-
-// CompositeStats is an internal type for collecting unifed realm and key server stats.
-type CompositeStats []*CompositeDay
-
-// CompositeData represents a single day of composite stats.
-type CompositeDay struct {
-	Day            time.Time
-	RealmStats     *database.RealmStat
-	KeyServerStats *keyserver.StatsDay
-}
-
-type jsonCompositeStat struct {
-	RealmID           uint                      `json:"realm_id"`
-	HasKeyServerStats bool                      `json:"has_key_server_stats"`
-	Stats             []*jsonCompositeStatStats `json:"statistics"`
-}
-
-type jsonCompositeStatStats struct {
-	Date time.Time                   `json:"date"`
-	Data *jsonCompositeStatStatsData `json:"data"`
-}
-
-type jsonCompositeStatStatsData struct {
-	database.JSONRealmStatStatsData
-	keyserver.StatsDay
-
-	// the only field that isn't from an embedded struct.
-	TotalPublishRequests int64 `json:"total_publish_requests"`
-}
-
-// MarshalJSON is a custom JSON marshaller.
-func (c CompositeStats) MarshalJSON() ([]byte, error) {
-	// Do nothing if there's no records
-	if len(c) == 0 {
-		return json.Marshal(struct{}{})
-	}
-	var realmID uint
-	hasKeyServerStats := false
-
-	stats := make([]*jsonCompositeStatStats, 0, len(c))
-	for _, stat := range c {
-		data := &jsonCompositeStatStatsData{}
-		if stat.RealmStats != nil {
-			if realmID == 0 {
-				realmID = stat.RealmStats.RealmID
-			}
-
-			data.CodesIssued = stat.RealmStats.CodesIssued
-			data.CodesClaimed = stat.RealmStats.CodesClaimed
-			data.CodesInvalid = stat.RealmStats.CodesInvalid
-			data.TokensClaimed = stat.RealmStats.TokensClaimed
-			data.TokensInvalid = stat.RealmStats.TokensInvalid
-			data.CodeClaimMeanAge = uint(stat.RealmStats.CodeClaimMeanAge.Duration.Seconds())
-			data.CodeClaimDistribution = stat.RealmStats.CodeClaimAgeDistribution
-		}
-		if stat.KeyServerStats != nil {
-			hasKeyServerStats = true
-			data.TotalPublishRequests = stat.KeyServerStats.PublishRequests.Total()
-			data.PublishRequests = stat.KeyServerStats.PublishRequests
-			data.TotalTEKsPublished = stat.KeyServerStats.TotalTEKsPublished
-			data.RevisionRequests = stat.KeyServerStats.RevisionRequests
-			data.TEKAgeDistribution = stat.KeyServerStats.TEKAgeDistribution
-			data.OnsetToUploadDistribution = stat.KeyServerStats.OnsetToUploadDistribution
-			data.RequestsMissingOnsetDate = stat.KeyServerStats.RequestsMissingOnsetDate
-		}
-
-		stats = append(stats, &jsonCompositeStatStats{
-			Date: stat.Day,
-			Data: data,
-		})
-	}
-
-	// Sort in descending order.
-	sort.Slice(stats, func(i, j int) bool {
-		return stats[i].Date.After(stats[j].Date)
-	})
-
-	var result jsonCompositeStat
-	result.RealmID = realmID
-	result.HasKeyServerStats = hasKeyServerStats
-	result.Stats = stats
-
-	b, err := json.Marshal(result)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal json: %w", err)
-	}
-	return b, nil
-}
-
-func (c *CompositeStats) UnmarshalJSON(b []byte) error {
-	if len(b) == 0 {
-		return nil
-	}
-	// This is unused, but required for the interface.
-	return nil
-}
-
-// TODO(mikehelmick) - Make this join part of the public API in key sever, then remove this
-func join(arr []int64, sep string) string {
-	var sb strings.Builder
-	for i, d := range arr {
-		sb.WriteString(strconv.FormatInt(d, 10))
-		if i != len(arr)-1 {
-			sb.WriteString(sep)
-		}
-	}
-	return sb.String()
-}
-
-// MarshalCSV returns bytes in CSV format.
-func (c CompositeStats) MarshalCSV() ([]byte, error) {
-	// Do nothing if there's no records
-	if len(c) == 0 {
-		return nil, nil
-	}
-
-	var b bytes.Buffer
-	w := csv.NewWriter(&b)
-
-	if err := w.Write([]string{
-		"date",
-		"codes_issued", "codes_claimed", "codes_invalid",
-		"tokens_claimed", "tokens_invalid", "code_claim_mean_age_seconds", "code_claim_age_distribution",
-		"publish_requests_unknown", "publish_requests_android", "publish_requests_ios",
-		"total_teks_published", "requests_with_revisions", "requests_missing_onset_date", "tek_age_distribution", "onset_to_upload_distribution",
-	}); err != nil {
-		return nil, fmt.Errorf("failed to write CSV header: %w", err)
-	}
-
-	// Due to the nature of the composite stats, one or the other of the branches could be nil
-	for i, stat := range c {
-		row := make([]string, 0, 16)
-		row = append(row, stat.Day.Format(project.RFC3339Date))
-		if stat.RealmStats == nil {
-			// no realm stats, 7 empty columns
-			row = append(row, "", "", "", "", "", "", "")
-		} else {
-			row = append(row, strconv.FormatUint(uint64(stat.RealmStats.CodesIssued), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.RealmStats.CodesClaimed), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.RealmStats.CodesInvalid), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.RealmStats.TokensClaimed), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.RealmStats.TokensInvalid), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.RealmStats.CodeClaimMeanAge.Duration.Seconds()), 10))
-			row = append(row, strings.Join(stat.RealmStats.CodeClaimAgeDistributionAsStrings(), "|"))
-		}
-		if stat.KeyServerStats == nil {
-			// no key sever stats, 8 empty columns
-			row = append(row, "", "", "", "", "", "", "", "")
-		} else {
-			row = append(row, strconv.FormatUint(uint64(stat.KeyServerStats.PublishRequests.UnknownPlatform), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.KeyServerStats.PublishRequests.Android), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.KeyServerStats.PublishRequests.IOS), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.KeyServerStats.TotalTEKsPublished), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.KeyServerStats.RevisionRequests), 10))
-			row = append(row, strconv.FormatUint(uint64(stat.KeyServerStats.RequestsMissingOnsetDate), 10))
-			row = append(row, join(stat.KeyServerStats.TEKAgeDistribution, "|"))
-			row = append(row, join(stat.KeyServerStats.OnsetToUploadDistribution, "|"))
-		}
-
-		if err := w.Write(row); err != nil {
-			return nil, fmt.Errorf("failed to write CSV entry %d: %w", i, err)
-		}
-	}
-
-	w.Flush()
-	if err := w.Error(); err != nil {
-		return nil, fmt.Errorf("failed to create CSV: %w", err)
-	}
-
-	return b.Bytes(), nil
-}
 
 // HandleComposite returns composite states for realm + key server
 // The key server stats may be omitted if that is not enabled
@@ -230,10 +43,10 @@ func (c *Controller) HandleComposite(typ Type) http.Handler {
 			return
 		}
 
-		var stats CompositeStats = make([]*CompositeDay, 0, len(realmStats))
-		statsMap := make(map[time.Time]*CompositeDay, len(realmStats))
+		stats := database.CompositeStats(make([]*database.CompositeDay, 0, len(realmStats)))
+		statsMap := make(map[time.Time]*database.CompositeDay, len(realmStats))
 		for _, rs := range realmStats {
-			day := &CompositeDay{
+			day := &database.CompositeDay{
 				Day:        rs.Date,
 				RealmStats: rs,
 			}
@@ -257,7 +70,7 @@ func (c *Controller) HandleComposite(typ Type) http.Handler {
 				compDay, ok := statsMap[ksDay.Day]
 				if !ok {
 					// if key server has stats from a day the realm doesn't, add it in.
-					compDay := &CompositeDay{
+					compDay := &database.CompositeDay{
 						Day: ksDay.Day,
 					}
 					stats = append(stats, compDay)
