+++ pkg/database/realm_stats_test.go
@@ -25,14 +25,16 @@ func TestRealmStats_MarshalCSV(t *testing.T) {
 	t.Parallel()
 
 	cases := []struct {
-		name  string
-		stats RealmStats
-		exp   string
+		name    string
+		stats   RealmStats
+		expCSV  string
+		expJSON string
 	}{
 		{
-			name:  "empty",
-			stats: nil,
-			exp:   "",
+			name:    "empty",
+			stats:   nil,
+			expCSV:  ``,
+			expJSON: `{}`,
 		},
 		{
 			name: "single",
@@ -49,9 +51,10 @@ func TestRealmStats_MarshalCSV(t *testing.T) {
 					CodeClaimAgeDistribution: []int32{1, 3, 4},
 				},
 			},
-			exp: `date,codes_issued,codes_claimed,codes_invalid,tokens_claimed,tokens_invalid,code_claim_mean_age_seconds,code_claim_age_distribution
+			expCSV: `date,codes_issued,codes_claimed,codes_invalid,tokens_claimed,tokens_invalid,code_claim_mean_age_seconds,code_claim_age_distribution
 2020-02-03,10,9,1,7,2,60,1|3|4
 `,
+			expJSON: `{"realm_id":1,"statistics":[{"date":"2020-02-03T00:00:00Z","data":{"codes_issued":10,"codes_claimed":9,"codes_invalid":1,"tokens_claimed":7,"tokens_invalid":2,"code_claim_mean_age_seconds":60,"code_claim_age_distribution":[1,3,4]}}]}`,
 		},
 		{
 			name: "multi",
@@ -90,11 +93,12 @@ func TestRealmStats_MarshalCSV(t *testing.T) {
 					CodeClaimAgeDistribution: []int32{7, 8, 9},
 				},
 			},
-			exp: `date,codes_issued,codes_claimed,codes_invalid,tokens_claimed,tokens_invalid,code_claim_mean_age_seconds,code_claim_age_distribution
+			expCSV: `date,codes_issued,codes_claimed,codes_invalid,tokens_claimed,tokens_invalid,code_claim_mean_age_seconds,code_claim_age_distribution
 2020-02-03,10,9,1,7,2,60,1|2|3
 2020-02-04,45,30,29,27,2,3600,4|5|6
 2020-02-05,15,2,0,2,0,0,7|8|9
 `,
+			expJSON: `{"realm_id":1,"statistics":[{"date":"2020-02-05T00:00:00Z","data":{"codes_issued":15,"codes_claimed":2,"codes_invalid":0,"tokens_claimed":2,"tokens_invalid":0,"code_claim_mean_age_seconds":0,"code_claim_age_distribution":[7,8,9]}},{"date":"2020-02-04T00:00:00Z","data":{"codes_issued":45,"codes_claimed":30,"codes_invalid":29,"tokens_claimed":27,"tokens_invalid":2,"code_claim_mean_age_seconds":3600,"code_claim_age_distribution":[4,5,6]}},{"date":"2020-02-03T00:00:00Z","data":{"codes_issued":10,"codes_claimed":9,"codes_invalid":1,"tokens_claimed":7,"tokens_invalid":2,"code_claim_mean_age_seconds":60,"code_claim_age_distribution":[1,2,3]}}]}`,
 		},
 	}
 
@@ -108,7 +112,7 @@ func TestRealmStats_MarshalCSV(t *testing.T) {
 			if err != nil {
 				t.Fatal(err)
 			}
-			if diff := cmp.Diff(string(b), tc.exp); diff != "" {
+			if diff := cmp.Diff(string(b), tc.expCSV); diff != "" {
 				t.Errorf("bad csv (+got, -want): %s", diff)
 			}
 
@@ -116,10 +120,8 @@ func TestRealmStats_MarshalCSV(t *testing.T) {
 			if err != nil {
 				t.Fatal(err)
 			}
-			a := &RealmStats{}
-			err = a.UnmarshalJSON(b)
-			if err != nil {
-				t.Fatal(err)
+			if got, want := string(b), tc.expJSON; got != want {
+				t.Errorf("bad json, expected \n%s\nto be\n%s\n", got, want)
 			}
 		})
 	}
